fixed rest controller so it returns http response

+++ src/main/java/fri/uni_lj/si/fileUploadService/api/FileController.java
@@ -2,13 +2,13 @@ package fri.uni_lj.si.fileUploadService.api;
 
 import fri.uni_lj.si.fileUploadService.models.FileData;
 import fri.uni_lj.si.fileUploadService.services.FileService;
-import fri.uni_lj.si.fileUploadService.services.FileStore;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.awt.*;
 import java.util.List;
 import java.util.UUID;
 
@@ -25,8 +25,9 @@ public class FileController {
     }
 
     @GetMapping
-    public List<FileData> getFiles() {
-        return fileService.getFiles();
+    public ResponseEntity<Object> getFiles() {
+        List<FileData> allFiles = fileService.getFiles();
+        return ResponseEntity.status(HttpStatus.OK).body(allFiles);
     }
 
     @PostMapping(
@@ -34,19 +35,41 @@ public class FileController {
             consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-
-    public FileData insertFileData (@RequestParam("file") MultipartFile file) {
+    public ResponseEntity<Object> insertFileData (@RequestParam("file") MultipartFile file) {
         FileData fd = new FileData(UUID.randomUUID(), file.getOriginalFilename(), "");
-        return fileService.insertFileData(fd, file);
+        FileData insertedFileData = fileService.insertFileData(fd, file);
+
+        if (insertedFileData.getTitle().equals("fail") && insertedFileData.getUri().equals("fail")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Failed to save data to DB.");
+        }
+        return ResponseEntity.status(HttpStatus.CREATED).body(insertedFileData);
     }
 
     @GetMapping(path = "{id}")
-    public FileData getFileDataById (@PathVariable("id") UUID id) {
-        return fileService.getFileDataById(id).orElse(null);
+    public ResponseEntity<Object> getFileDataById (@PathVariable("id") UUID id) {
+        if (id.toString().equals("")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Id of file is a required param");
+        }
+
+        FileData fd = fileService.getFileDataById(id).orElse(null);
+
+        if (fd == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+        }
+        return ResponseEntity.ok(fd);
     }
 
     @DeleteMapping(path = "{id}")
-    public void deleteFileDataById (@PathVariable("id") UUID id) {
-        fileService.deleteFileDataById(id);
+    public ResponseEntity<Object> deleteFileDataById (@PathVariable("id") UUID id) {
+        if (id.toString().equals("")) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Id of file is a required param");
+        }
+
+        int isDeleted = fileService.deleteFileDataById(id);
+
+        if (isDeleted == 1) {
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("FileData with id: " + id + " was succesfully deleted.");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("FileData with id: " + id + " was not found.");
     }
 }
